instance/requirements.txt
aiogram
python-dotenv
aioredis